move full text search filter related classes to one package
https://github.com/metasfresh/metasfresh/issues/5101
