+++ src/main/java/de/metas/ui/web/document/filter/provider/fullTextSearch/FullTextSearchDocumentFilterDescriptorsProviderFactory.java
@@ -21,8 +21,6 @@ import de.metas.ui.web.document.filter.provider.ImmutableDocumentFilterDescripto
 import de.metas.ui.web.document.filter.provider.NullDocumentFilterDescriptorsProvider;
 import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
-import de.metas.ui.web.window.descriptor.FullTextSearchFilterContext;
-import de.metas.ui.web.window.descriptor.FullTextSearchSqlDocumentFilterConverter;
 import de.metas.util.Services;
 
 /*
